Fixed a bug in z-ordering for `OverlayMode`
The stable ordering of windows introduced a bug in the z-ordering of
windows in `OverlayMode`, since its thumbnails are dependent on the
order the system originally returns window references when iterating
through them.

Now, the stable ordering will just be used to generate the according
window jump codes, but not for the z-ordering of the thumbnails.