Disable DPI awareness for WPF
Without this switch, the monitor overlays will get placed at weird
offsets if Sextant is run in an environment with screens using different
DPI scaling factor.

Since I don't have a monitor with very high DPI, I can't really tell if
this approach would be problematic in that case; but since the overlay
icons and fonts are quite large anyway, it seems good enough for now.